Adds nearly fully fledged DMs
But needed a good wrangling to get code out of the ChannelModel.

The abstraction I want to use is Server/Channel/User, but IRC
(both the protocol and library implementation) has only one
interface for all of these, which is "talk to the server", so I've
moved anything that talks to the client variable into the
ServerModel in the hopes that it keeps things more under control
without having to pass a lot of information around.

This was very nearly an architectural deadend, but getting past
it took less than 15 minutes to copy/paste/debug everything in
ChannelModel that used the IRC client parameter passed in after
I'd moved it to ServerModel.

I'm seeing a lot of repeated code in the ServerModel, so at some
point some refactoring will need to be undertaken to tidy some of
that up. I'm not hugely keen on the fact that I have to control a
the channel unread state from outside the ChannelModel itself, but
it was messy trying to do that from inside any of the other
handlers.