Messages in channels can now be saved, and are reloaded on startup
Unfortunately, this is via the /save command in the channel you
want to store.

I can hook this up to client-quit fairly trivially, but my current
development cycle currently consists of ctrl+S, alt-tab, ctrl+R,
and I'm struggling to trap the ctrl+R event in Electron.

Documentation (and SO) claim that the will-navigate event from
webContents should notice a change in the page, enough to trap
the event and do something with it, but so far I've had no success.