Channel messages are saved automatically on reload/quit
The reload/quit event trapping turned out to be the easiest
part of this whole problem (although I don't fully understand
what I've written, it seems to work ... ).

The part that was much harder, was making system messages
appear italic on reload.

What I want is:

messages = observableArray([
  {
    from: ...,
    fragments: observableArray([
      {
        message: ...,
        type: ...
      },
     ...
  }
])

This way, any time the messages array is updated (eg a new message
arrives) or the fragments array within an existing message is
updated (eg a new message has been appended), the screen will
update as appropriate.

Unfortunately, using the mapping plugin will unserialise this into:

messages = oA([
  {
    from: o(),
    fragments: oA([
      {
        message: o(),
        type: o()
       }
     ...
   }
])

Turning *everything* into an observable of some variety, including
all of the keys in the fragments array.

This wouldn't be an issue, except we use fragments.type to identify
system messages to the template (via CSS), using something like:

  css: {'system_message': $data.type == 'system'}

And that implied 'if' statement fails if type is an observable,
since it'll pop back as a function.

The messages.from isn't a problem, because knockoutjs coerces the
html value (data-bind="html: from") out of the function correctly
without complaining.

So what do we do? Well, we could scrap using the mapping plugin
entirely, and essentially do manual SerDes on the messages array,
or just turn the fragment information into (unnecessary)
observables and update the template to dereference the function
to give us the value back straight.

The latter is sucky, but it's what we've got right now. There is
probably a better answer, but performance isn't a problem _right
now_, so let's move on until it is.