Make links clickable
Using the linkify library, we parse incoming messages to add
appropriate <a> tags.

What this doesn't do is move the link target to an actual browser,
which is pretty bad right now. The answer is somewhere in:
  require("shell").openExternal("http://www.google.com")

Services like embed.ly and iframely.com provide the "cards"
interface I'd want to use, but are best done server-side, since
API keys and caching would be desirable here.

A hacky version might be possible with
  http://html2canvas.hertzen.com/documentation.html or similar.